# backend/src/main/java/andrearepellini/phonebook/dto/response/LoginResponse.java
package andrearepellini.phonebook.dto.response;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class LoginResponse {
    private String token;
    private long expiresIn;
}