# schema/openapi.yaml
openapi: 3.1.0
info:
  title: Phonebook API
  description: Backend API for managing phonebook contacts
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Contacts
  description: Contact management APIs
- name: Auth
  description: User authentication APIs
- name: Hello world
  description: Hello world test API
paths:
  /api/contacts:
    get:
      tags:
      - Contacts
      summary: Get all contacts
      description: Retrieve a paginated list of contacts with optional filtering
      operationId: getAllContacts
      parameters:
      - name: filter
        in: query
        required: false
        schema:
          type: string
      - name: deleted
        in: query
        required: false
        schema:
          type: boolean
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelContactResponse"
    post:
      tags:
      - Contacts
      summary: Create a new contact
      description: Create a new contact with the provided details
      operationId: createContact
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
        required: true
      responses:
        "201":
          description: Contact created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
  /api/auth/signup:
    post:
      tags:
      - Auth
      summary: Register a new user
      description: Creates a new user account with the provided registration details
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
        required: true
      responses:
        "200":
          description: User successfully registered
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid input or email already registered
          content:
            '*/*':
              schema:
                type: string
  /api/auth/login:
    post:
      tags:
      - Auth
      summary: Authenticate a user
      description: Authenticates a user using their credentials and returns a JWT
        token for access
      operationId: authenticateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthenticateUserRequest"
        required: true
      responses:
        "200":
          description: Successfully authenticated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid input or user not registered
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Invalid credentials
          content:
            '*/*':
              schema:
                type: string
  /api/contacts/{id}:
    get:
      tags:
      - Contacts
      summary: Get contact by ID
      description: Retrieve a single contact by its ID
      operationId: getContactById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Contact found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "404":
          description: Contact not found
    patch:
      tags:
      - Contacts
      summary: Update a contact
      description: Update an existing contact's details
      operationId: patchContact
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchContactRequest"
        required: true
      responses:
        "200":
          description: Contact updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "404":
          description: Contact not found
  /api/helloworld:
    get:
      tags:
      - Hello world
      operationId: helloWorld
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
components:
  schemas:
    CreateContactRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        address:
          type: string
        phoneNumber:
          type: string
        age:
          type: integer
          format: int32
    ContactResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        address:
          type: string
        phoneNumber:
          type: string
        age:
          type: integer
          format: int32
        deleted:
          type: boolean
    RegisterUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          maxLength: 2147483647
          minLength: 8
      required:
      - email
      - password
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
    AuthenticateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
      - email
      - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        expiresIn:
          type: integer
          format: int64
    PatchContactRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        address:
          type: string
        phoneNumber:
          type: string
        age:
          type: integer
          format: int32
        deleted:
          type: boolean
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModelContactResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ContactResponse"
        page:
          $ref: "#/components/schemas/PageMetadata"